change some static fields to dynamic, move Jmccc back
