+++ src/main/java/com/darkyoooooo/jmccc/launch/Reporter.java
@@ -15,7 +15,7 @@ import com.google.gson.Gson;
  * The extended identity is used to identity the caller of JMCCC, default to null. If you wanna help us do the
  * statistics better, please set this to the name, and the version of your launcher.
  */
-class Reporter {
+public class Reporter {
 
     /**
      * JMCCC Report API
